+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/AbstractBaseScreen.java
@@ -9,6 +9,7 @@ import com.refinedmods.refinedstorage2.platform.common.screen.widget.ResourceFil
 
 import java.util.List;
 import java.util.Objects;
+import javax.annotation.Nullable;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
@@ -216,5 +217,10 @@ public abstract class AbstractBaseScreen<T extends AbstractContainerMenu> extend
             minecraft.setScreen(new ResourceAmountScreen(this, playerInventory, ((ResourceFilterSlot) slot)));
         }
         return canChangeAmount;
+    }
+
+    @Nullable
+    public Slot getHoveredSlot() {
+        return this.hoveredSlot;
     }
 }