+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/grid/AbstractGridScreen.java
@@ -408,6 +408,14 @@ public abstract class AbstractGridScreen<R, T extends AbstractGridContainerMenu<
             && lastModified.amount() <= MODIFIED_JUST_NOW_MAX_SECONDS;
     }
 
+    @Nullable
+    public AbstractGridResource<?> getCurrentlyHoveredResource() {
+        if (this.gridSlotNumber == -1) {
+            return null;
+        }
+        return menu.getView().getAll().get(this.gridSlotNumber);
+    }
+
     @Override
     public void render(final PoseStack poseStack, final int mouseX, final int mouseY, final float partialTicks) {
         super.render(poseStack, mouseX, mouseY, partialTicks);