+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed missing recoloring recipes for Grid and Controller to default color.
+-   Fixed missing recoloring recipes for Fluid Grid.
+
 ## [v2.0.0-milestone.2.5] - 2023-01-11
 
 ### Fixed